+++ Utility/Random.h
@@ -10,7 +10,7 @@ namespace SnackerEngine
 
 	float randomFloat(float min = std::numeric_limits<float>::min(), float max = std::numeric_limits<float>::max());
 
-	float randomDouble(double min = std::numeric_limits<double>::min(), double max = std::numeric_limits<double>::max());
+	double randomDouble(double min = std::numeric_limits<double>::min(), double max = std::numeric_limits<double>::max());
 
 	/// Returns a random integer in [min, max]
 	int randomInteger(int min, int max);
